+++ udata/core/dataservices/tasks.py
@@ -1,8 +1,7 @@
 from celery.utils.log import get_task_logger
 
 from udata.core.dataservices.models import Dataservice
-
-# from udata.harvest.models import HarvestJob
+from udata.harvest.models import HarvestJob
 from udata.models import Discussion, Follow, Transfer
 from udata.tasks import job
 
@@ -18,8 +17,7 @@ def purge_dataservices(self):
         # Remove discussions
         Discussion.objects(subject=dataservice).delete()
         # Remove HarvestItem references
-        # TODO: uncomment when adding dataservice harvest
-        # HarvestJob.objects(items__dataservice=dataservice).update(set__items__S__dataservice=None)
+        HarvestJob.objects(items__dataservice=dataservice).update(set__items__S__dataservice=None)
         # Remove associated Transfers
         Transfer.objects(subject=dataservice).delete()
         # Remove dataservice