+++ udata/tests/dataset/test_dataset_tasks.py
@@ -1,17 +1,20 @@
 import pytest
 
-from udata.core.dataset import tasks
-
 # Those imports seem mandatory for the csv adapters to be registered.
 # This might be because of the decorator mechanism.
+from udata.core.dataservices.models import Dataservice
+from udata.core.dataset import tasks
 from udata.core.dataset.csv import DatasetCsvAdapter, ResourcesCsvAdapter  # noqa
 from udata.core.dataset.factories import CommunityResourceFactory, DatasetFactory
+from udata.core.discussions.factories import DiscussionFactory
 from udata.core.organization.csv import OrganizationCsvAdapter  # noqa
 from udata.core.reuse.csv import ReuseCsvAdapter  # noqa
 from udata.core.tags.csv import TagCsvAdapter  # noqa
 from udata.core.user.factories import UserFactory
 from udata.harvest.csv import HarvestSourceCsvAdapter  # noqa
-from udata.models import CommunityResource, Dataset, Topic, Transfer
+from udata.harvest.models import HarvestItem, HarvestJob
+from udata.harvest.tests.factories import HarvestJobFactory
+from udata.models import CommunityResource, Dataset, Discussion, Follow, Topic, Transfer
 
 pytestmark = pytest.mark.usefixtures("clean_db")
 
@@ -32,12 +35,25 @@ def test_purge_datasets():
         comment="comment",
     )
 
+    discussion = DiscussionFactory(subject=datasets[0])
+
+    follower = Follow.objects.create(follower=user, following=datasets[0])
+
+    HarvestJobFactory(items=[HarvestItem(dataset=datasets[0])])
+
+    Dataservice.objects.create(title="test", datasets=datasets)
+
     tasks.purge_datasets()
 
     assert Transfer.objects.filter(id=transfer.id).count() == 0
 
     topic = Topic.objects(name="test topic").first()
     assert topic.datasets[0] == datasets[1]
+
+    assert Discussion.objects.filter(id=discussion.id).count() == 0
+    assert Follow.objects.filter(id=follower.id).count() == 0
+    assert HarvestJob.objects.filter(items__dataset=datasets[0].id).count() == 0
+    assert Dataservice.objects.filter(datasets=datasets[0].id).count() == 0
 
 
 def test_purge_datasets_community():